requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.8.0
setuptools==65.5.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
pillow